requirements..txt
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow
qrcode
fpdf
requests
python-barcode
google-genai
gunicorn